# lib/api-spec/openapi.yaml
...
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
paths:
...
  /projects:
    get:
      tags:
      - projects
      summary: List Projects
      operationId: getProjects
      parameters:
      - name: studyId
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Studyid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectOut'
                title: Response Getprojects
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /dqa/summary:
    get:
      tags:
      - dqa
      summary: Dqa Summary
      operationId: getDqaSummary
      parameters:
      - name: studyId
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Studyid
      - name: projectId
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Projectid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DqaSummary'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /dqa/by-project:
    get:
      tags:
      - dqa
      summary: Dqa By Project
      description: Per-project DQA metrics aligned with the Data Quality dashboard.
      operationId: getDqaByProject
      parameters:
      - name: studyId
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Studyid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectDqaStat'
                title: Response Getdqabyproject
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /dqa/flags:
    get:
      tags:
      - dqa
      summary: List Flags
      operationId: getDqaFlags
      parameters:
      - name: projectId
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Projectid
      - name: studyId
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Studyid
      - name: submissionId
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Submissionid
      - name: ruleId
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Ruleid
      - name: severity
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Severity
      - name: enumerator
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Enumerator
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 2000
          minimum: 1
          default: 500
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DqaFlagOut'
                title: Response Getdqaflags
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /dqa/enumerators:
    get:
      tags:
      - dqa
      summary: Enumerator Stats
      operationId: getDqaEnumerators
      parameters:
      - name: studyId
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Studyid
      - name: projectId
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Projectid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/app__schemas__dqa__EnumeratorStat'
                title: Response Getdqaenumerators
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /dqa/recompute:
    post:
      tags:
      - dqa
      summary: Recompute
      operationId: recomputeDqa
      parameters:
      - name: projectId
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Projectid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DqaRecomputeResult'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /dqa/triangulation/{view_id}:
    get:
      tags:
      - dqa
      summary: Triangulation View
      description: 'UDISE-joined triangulation views: TR-1, TR-3, TR-5 (study-scoped
        when studyId set).'
      operationId: getTriangulationView
      parameters:
      - name: view_id
        in: path
        required: true
        schema:
          type: string
          title: View Id
      - name: studyId
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Studyid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TriangulationViewOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /projects/{project_id}/form-fields:
    get:
      tags:
      - dqa
      summary: Project Form Fields
      operationId: getProjectFormFields
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          title: Project Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FormFieldOut'
                title: Response Getprojectformfields
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /projects/{project_id}/rule-pack:
    get:
      tags:
      - dqa
      summary: Get Rule Pack
      operationId: getProjectRulePack
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          title: Project Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RulePackOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - dqa
      summary: Put Rule Pack
      operationId: updateProjectRulePack
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          title: Project Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RulePackUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RulePackOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
...
    DqaFlagOut:
      properties:
        id:
          type: string
          title: Id
        submissionId:
          type: string
          title: Submissionid
        projectId:
          type: string
          title: Projectid
        ruleId:
          type: string
          title: Ruleid
        severity:
          type: string
          title: Severity
        title:
          type: string
          title: Title
        message:
          type: string
          title: Message
        details:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Details
        evaluatedAt:
          type: string
          title: Evaluatedat
        enumerator:
          anyOf:
          - type: string
          - type: 'null'
          title: Enumerator
        projectName:
          anyOf:
          - type: string
          - type: 'null'
          title: Projectname
        koboId:
          anyOf:
          - type: string
          - type: 'null'
          title: Koboid
        submittedAt:
          anyOf:
          - type: string
          - type: 'null'
          title: Submittedat
      type: object
      required:
      - id
      - submissionId
      - projectId
      - ruleId
      - severity
      - title
      - message
      - evaluatedAt
      title: DqaFlagOut
    DqaRecomputeResult:
      properties:
        projectId:
          anyOf:
          - type: string
          - type: 'null'
          title: Projectid
        submissions:
          type: integer
          title: Submissions
        flaggedSubmissions:
          type: integer
          title: Flaggedsubmissions
        flags:
          type: integer
          title: Flags
      type: object
      required:
      - submissions
      - flaggedSubmissions
      - flags
      title: DqaRecomputeResult
...
    DqaSummary:
      properties:
        projectId:
          anyOf:
          - type: string
          - type: 'null'
          title: Projectid
        totalSubmissions:
          type: integer
          title: Totalsubmissions
        flaggedSubmissions:
          type: integer
          title: Flaggedsubmissions
        flaggedPct:
          type: number
          title: Flaggedpct
        redFlags:
          type: integer
          title: Redflags
        amberFlags:
          type: integer
          title: Amberflags
        byRule:
          items:
            $ref: '#/components/schemas/DqaRuleCount'
          type: array
          title: Byrule
      type: object
      required:
      - totalSubmissions
      - flaggedSubmissions
      - flaggedPct
      - redFlags
      - amberFlags
      - byRule
      title: DqaSummary
...
    FormFieldOut:
      properties:
        name:
          type: string
          title: Name
        type:
          type: string
          title: Type
        label:
          type: string
          title: Label
        listName:
          anyOf:
          - type: string
          - type: 'null'
          title: Listname
        choices:
          items:
            additionalProperties: true
            type: object
          type: array
          title: Choices
          default: []
      type: object
      required:
      - name
      - type
      - label
      title: FormFieldOut
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ProjectDqaStat:
      properties:
        projectId:
          type: string
          title: Projectid
        projectName:
          type: string
          title: Projectname
        totalSubmissions:
          type: integer
          title: Totalsubmissions
        cleanSubmissions:
          type: integer
          title: Cleansubmissions
        amberSubmissions:
          type: integer
          title: Ambersubmissions
        redSubmissions:
          type: integer
          title: Redsubmissions
        redFlags:
          type: integer
          title: Redflags
        amberFlags:
          type: integer
          title: Amberflags
        flaggedPct:
          type: number
          title: Flaggedpct
      type: object
      required:
      - projectId
      - projectName
      - totalSubmissions
      - cleanSubmissions
      - amberSubmissions
      - redSubmissions
      - redFlags
      - amberFlags
      - flaggedPct
      title: ProjectDqaStat
    ProjectOut:
      properties:
        id:
          type: string
          title: Id
        uid:
          type: string
          title: Uid
        name:
          type: string
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        status:
          type: string
          title: Status
        submissionCount:
          type: integer
          title: Submissioncount
        formCount:
          type: integer
          title: Formcount
        enumeratorCount:
          type: integer
          title: Enumeratorcount
        questionCount:
          type: integer
          title: Questioncount
        owner:
          anyOf:
          - type: string
          - type: 'null'
          title: Owner
        lastEditedBy:
          anyOf:
          - type: string
          - type: 'null'
          title: Lasteditedby
        lastModifiedAt:
          anyOf:
          - type: string
          - type: 'null'
          title: Lastmodifiedat
        lastSubmissionAt:
          anyOf:
          - type: string
          - type: 'null'
          title: Lastsubmissionat
        lastSyncAt:
          anyOf:
          - type: string
          - type: 'null'
          title: Lastsyncat
        createdAt:
          type: string
          title: Createdat
        deployedAt:
          anyOf:
          - type: string
          - type: 'null'
          title: Deployedat
        formVersionId:
          anyOf:
          - type: string
          - type: 'null'
          title: Formversionid
        deployedVersionId:
          anyOf:
          - type: string
          - type: 'null'
          title: Deployedversionid
        currentVersionNumber:
          anyOf:
          - type: integer
          - type: 'null'
          title: Currentversionnumber
        deployedVersionNumber:
          anyOf:
          - type: integer
          - type: 'null'
          title: Deployedversionnumber
        syncStatus:
          type: string
          title: Syncstatus
        syncError:
          anyOf:
          - type: string
          - type: 'null'
          title: Syncerror
        sector:
          anyOf:
          - type: string
          - type: 'null'
          title: Sector
        country:
          anyOf:
          - type: string
          - type: 'null'
          title: Country
        labelLanguage:
          type: string
          title: Labellanguage
        availableLabelLanguages:
          items:
            type: string
          type: array
          title: Availablelabellanguages
        studyId:
          anyOf:
          - type: string
          - type: 'null'
          title: Studyid
        toolCode:
          anyOf:
          - type: string
          - type: 'null'
          title: Toolcode
        studyName:
          anyOf:
          - type: string
          - type: 'null'
          title: Studyname
      type: object
      required:
      - id
      - uid
      - name
      - status
      - submissionCount
      - formCount
      - enumeratorCount
      - questionCount
      - createdAt
      - syncStatus
      - labelLanguage
      - availableLabelLanguages
      title: ProjectOut
...
    TriangulationViewOut:
      properties:
        id:
          type: string
          title: Id
        title:
          type: string
          title: Title
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        rows:
          items:
            $ref: '#/components/schemas/TriangulationRow'
          type: array
          title: Rows
        mismatchCount:
          type: integer
          title: Mismatchcount
        practices:
          items:
            $ref: '#/components/schemas/TriangulationPracticeStat'
          type: array
          title: Practices
          default: []
      type: object
      required:
      - id
      - title
      - rows
      - mismatchCount
      title: TriangulationViewOut
...
    app__schemas__dqa__EnumeratorStat:
      properties:
        enumerator:
          type: string
          title: Enumerator
        submissions:
          type: integer
          title: Submissions
        flagged:
          type: integer
          title: Flagged
        flaggedPct:
          type: number
          title: Flaggedpct
        redFlags:
          type: integer
          title: Redflags
        amberFlags:
          type: integer
          title: Amberflags
          default: 0
        medianDurationMinutes:
          anyOf:
          - type: number
          - type: 'null'
          title: Mediandurationminutes
      type: object
      required:
      - enumerator
      - submissions
      - flagged
      - flaggedPct
      - redFlags
      title: EnumeratorStat
...
servers:
- url: /api
  description: Base API path